Creating interfaces for each step
