+++ gen/config/validate.go
@@ -1,0 +1,63 @@
+package config
+
+import (
+	"fmt"
+	"log"
+	"os"
+
+	"github.com/vtex/action-cm-generator/gen"
+	"github.com/vtex/action-cm-generator/gen/util"
+	"github.com/xeipuuv/gojsonschema"
+)
+
+// Validator is responsible for validate configuration contents.
+type Validator struct{}
+
+// Validate returns a channel of all validated files.
+func (v *Validator) Validate(in <-chan gen.Config) (out <-chan gen.Config) {
+	ch := make(chan gen.Config)
+	logger := log.New(os.Stdout, log.Prefix(), log.Flags())
+
+	go func() {
+		for config := range in {
+			logger.SetPrefix("[validator]: ")
+
+			ls := gojsonschema.NewGoLoader(config.Schema)
+			cl := gojsonschema.NewGoLoader(config.Content)
+			result, err := gojsonschema.Validate(ls, cl)
+
+			if err != nil {
+				schema, err := util.MarshalIndent(config.Schema)
+
+				if err != nil {
+					logger.Println(err)
+				}
+
+				content, err := util.MarshalIndent(config.Content)
+
+				if err != nil {
+					logger.Println(err)
+				}
+
+				logger.Fatal(fmt.Errorf("\n>>>>>> Schema\n %s\n>>>>>> Content\n %s\n error when trying to validate the config %s %v", schema, content, config.Path, err))
+			}
+
+			resultErrs := result.Errors()
+
+			if len(resultErrs) > 0 || !result.Valid() {
+				logger.Fatalln(resultErrs)
+			}
+
+			ch <- config
+		}
+
+		close(ch)
+	}()
+
+	return ch
+}
+
+// NewValidator creates a new validator instance.
+func NewValidator() *Validator {
+	return &Validator{}
+}
